auth_service/app/requirements.txt
fastapi==0.95.2
uvicorn==0.21.1
passlib[bcrypt]==1.7.4
psycopg2==2.9.6
PyJWT==2.7.0